requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.4.0
certifi==2022.6.15
chardet==4.0.0
discord.py==1.7.3
idna==3.3
multidict==6.0.2
typing-extensions==4.2.0
yarl==1.7.2
